Add --enable-debug to configure script to fix CFLAGS + CPPFLAGS
This patch adds an --enable-debug flag to the configure script
to control the CPPFLAGS and CFLAGS at built time.

In debug mode we disable all optimization flags and enable -g to
get GDB symbols in the binary.  We also define DEBUG=1 so that
the LOG() macro sends all debug messages to the gul.log file.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>